requirements.txt
beautifulsoup4==4.9.3
pylint==2.6.0
redis==3.5.3
requests==2.25.0
nltk~=3.5
PyYAML~=5.3.1
elasticsearch~=7.10.0
setuptools~=50.3.2